refactor: split library into modules
This commit moves some of the structs and interfaces
for `library` into their own files, this change is
made in preparation for new `deps` interface
refactoring.

No functional changes have been made.
