+++ Blueprints
@@ -40,20 +40,26 @@ bootstrap_go_package {
         "bob-depmap",
     ],
     srcs: [
+        "core/alias.go",
         "core/android.go",
         "core/androidbp_backend.go",
         "core/androidbp_cclibs.go",
-        "core/androidbp_kernel_module.go",
-        "core/androidbp_resource.go",
         "core/androidbp_filegroup.go",
         "core/androidbp_generated.go",
-        "core/alias.go",
+        "core/androidbp_kernel_module.go",
+        "core/androidbp_resource.go",
+        "core/binary.go",
+        "core/build.go",
+        "core/build_props.go",
         "core/build_structs.go",
+        "core/common_props.go",
         "core/config_props.go",
         "core/defaults.go",
-        "core/external_library.go",
+        "core/dep_sorter.go",
         "core/escape.go",
+        "core/external_library.go",
         "core/feature.go",
+        "core/filegroup.go",
         "core/filepath.go",
         "core/gen_binary.go",
         "core/gen_library.go",
@@ -63,30 +69,33 @@ bootstrap_go_package {
         "core/glob.go",
         "core/graphviz.go",
         "core/install.go",
-        "core/filegroup.go",
         "core/kernel_module.go",
         "core/late_template.go",
-        "core/library.go",
-        "core/strict_library.go",
-        "core/source_props.go",
         "core/legacy_source_props.go",
+        "core/library.go",
+        "core/library_shared.go",
+        "core/library_static.go",
+        "core/linux_backend.go",
+        "core/linux_cclibs.go",
+        "core/linux_generated.go",
+        "core/linux_kernel_module.go",
         "core/output_producer.go",
         "core/properties.go",
+        "core/property_exporter.go",
+        "core/source_props.go",
         "core/splitter.go",
         "core/standalone.go",
+        "core/strict_library.go",
+        "core/strict_linux_cclibs.go",
         "core/strip.go",
         "core/template.go",
         "core/toolchain.go",
-        "core/linux_backend.go",
-        "core/linux_cclibs.go",
-        "core/strict_linux_cclibs.go",
-        "core/linux_generated.go",
-        "core/linux_kernel_module.go",
     ],
     testSrcs: [
+        "core/android_test.go",
+        "core/androidbp_test.go",
         "core/feature_test.go",
         "core/template_test.go",
-        "core/androidbp_test.go",
     ],
     pkgPath: "github.com/ARM-software/bob-build/core",
 }
