+++ core/BUILD.bazel
@@ -11,9 +11,14 @@ go_library(
         "androidbp_generated.go",
         "androidbp_kernel_module.go",
         "androidbp_resource.go",
+        "binary.go",
+        "build.go",
+        "build_props.go",
         "build_structs.go",
+        "common_props.go",
         "config_props.go",
         "defaults.go",
+        "dep_sorter.go",
         "escape.go",
         "external_library.go",
         "feature.go",
@@ -31,12 +36,15 @@ go_library(
         "late_template.go",
         "legacy_source_props.go",
         "library.go",
+        "library_shared.go",
+        "library_static.go",
         "linux_backend.go",
         "linux_cclibs.go",
         "linux_generated.go",
         "linux_kernel_module.go",
         "output_producer.go",
         "properties.go",
+        "property_exporter.go",
         "source_props.go",
         "splitter.go",
         "standalone.go",
