+++ core/library.go
@@ -22,274 +22,13 @@ import (
 	"fmt"
 	"path/filepath"
 	"regexp"
-	"strings"
 
 	"github.com/google/blueprint"
 
-	"github.com/ARM-software/bob-build/internal/graph"
 	"github.com/ARM-software/bob-build/internal/utils"
 )
 
-const (
-	tocExt = ".toc"
-)
-
 var depOutputsVarRegexp = regexp.MustCompile(`^\$\{(.+)_out\}$`)
-
-type propertyExporter interface {
-	exportCflags() []string
-	exportIncludeDirs() []string
-	exportSystemIncludeDirs() []string
-	exportLdflags() []string
-	exportLdlibs() []string
-	exportLocalIncludeDirs() []string
-	exportLocalSystemIncludeDirs() []string
-	exportSharedLibs() []string
-}
-
-// CommonProps defines a set of properties which are common
-// for multiple module types.
-type CommonProps struct {
-	LegacySourceProps
-	IncludeDirsProps
-	InstallableProps
-	EnableableProps
-	AndroidProps
-	AliasableProps
-
-	// Flags used for C compilation
-	Cflags []string
-}
-
-// BuildProps contains properties required by all modules that compile C/C++
-type BuildProps struct {
-	// Alternate output name, used for the file name and Android rules
-	Out *string
-	// Flags exported for dependent modules
-	Export_cflags []string
-	// Flags used for C compilation
-	Conlyflags []string
-	// Flags used for C++ compilation
-	Cxxflags []string
-	// Flags used for assembly compilation
-	Asflags []string
-	// Flags used for linking
-	Ldflags []string
-	// Same as ldflags, but specified on static libraries and propagated to
-	// the top-level build object.
-	Export_ldflags []string
-	// Shared library version
-	Library_version string
-	// Shared library version script
-	Version_script *string
-
-	// The list of shared lib modules that this library depends on.
-	// These are propagated to the closest linking object when specified on static libraries.
-	// shared_libs is an indication that this module is using a shared library, and
-	// users of this module need to link against it.
-	Shared_libs []string `bob:"first_overrides"`
-	// The libraries mentioned here will be appended to shared_libs of the modules that use
-	// this library (via static_libs, whole_static_libs or shared_libs).
-	ExtraSharedLibs []string `blueprint:"mutated"`
-
-	// The list of static lib modules that this library depends on
-	// These are propagated to the closest linking object when specified on static libraries.
-	// static_libs is an indication that this module is using a static library, and
-	// users of this module need to link against it.
-	Static_libs []string `bob:"first_overrides"`
-
-	// This list of dependencies that exported cflags and exported include dirs
-	// should be propagated 1-level higher
-	Reexport_libs []string `bob:"first_overrides"`
-	// Internal property for collecting libraries with reexported flags and include paths
-	ResolvedReexportedLibs []string `blueprint:"mutated"`
-
-	ResolvedStaticLibs []string `blueprint:"mutated"`
-
-	// The list of whole static libraries that this library depnds on
-	// This will include all the objects in the library (as opposed to normal static linking)
-	// If this is set for a static library, any shared library will also include objects
-	// from dependent libraries
-	Whole_static_libs []string `bob:"first_overrides"`
-
-	// List of libraries to import headers from, but not link to
-	Header_libs []string `bob:"first_overrides"`
-
-	// List of libraries that users of the current library should import
-	// headers from, but not link to
-	Export_header_libs []string `bob:"first_overrides"`
-
-	// Linker flags required to link to the necessary system libraries
-	// These are propagated to the closest linking object when specified on static libraries.
-	Ldlibs []string `bob:"first_overrides"`
-
-	// The list of modules that generate extra headers for this module
-	Generated_headers []string `bob:"first_overrides"`
-
-	// The list of modules that generate extra headers for this module,
-	// which should be made available to linking modules
-	Export_generated_headers []string `bob:"first_overrides"`
-
-	// The list of modules that generate extra source files for this module
-	Generated_sources []string
-
-	// The list of modules that generate output required by the build wrapper
-	Generated_deps []string
-
-	// Include local dirs to be exported into dependent.
-	// The system variant will propagate includes using `-isystem`, but use `-I` for
-	// current module.
-	Export_local_include_dirs        []string `bob:"first_overrides"`
-	Export_local_system_include_dirs []string `bob:"first_overrides"`
-
-	// Include dirs (path relative to root) to be exported into dependent.
-	// The system variant will propagate includes using `-isystem`, but use `-I` for
-	// current module.
-	Export_include_dirs        []string `bob:"first_overrides"`
-	Export_system_include_dirs []string `bob:"first_overrides"`
-
-	// Wrapper for all build commands (object file compilation *and* linking)
-	Build_wrapper *string
-
-	// Adds DT_RPATH symbol to binaries and shared libraries so that they can find
-	// their dependencies at runtime.
-	Add_lib_dirs_to_rpath *bool
-
-	// This is a shared library that pulls in one or more shared
-	// libraries to resolve symbols that the binary needs. This is
-	// useful where a named library is the standard library to link
-	// against, but the implementation may exist in another
-	// library.
-	//
-	// Only valid on bob_shared_library.
-	//
-	// Currently we need to link with -Wl,--copy-dt-needed-entries.
-	// This makes the binary depend on the implementation library, and
-	// requires the BFD linker.
-	Forwarding_shlib *bool
-
-	StripProps
-	AndroidPGOProps
-	AndroidMTEProps
-
-	Hwasan_enabled *bool
-
-	TargetType tgtType `blueprint:"mutated"`
-}
-
-func (b *BuildProps) processBuildWrapper(ctx blueprint.BaseModuleContext) {
-	if b.Build_wrapper != nil {
-		// The build wrapper may be a local tool, in which case we
-		// need to prefix it with ${SrcDir}. It can also be a tool in
-		// PATH like ccache.
-		//
-		// We want to avoid doing this repeatedly, so try do it in an
-		// early mutator
-		*b.Build_wrapper = strings.TrimSpace(*b.Build_wrapper)
-		firstWord := strings.SplitN(*b.Build_wrapper, " ", 1)[0]
-
-		// If the first character is '/' this is an absolute path, so no need to do anything
-		if firstWord[0] != '/' {
-			// Otherwise if the first word contains '/' this is a local path
-			if strings.ContainsAny(firstWord, "/") {
-				*b.Build_wrapper = getBackendPathInSourceDir(getBackend(ctx), *b.Build_wrapper)
-			}
-		}
-	}
-}
-
-// A Build represents the whole tree of properties for a 'library' object,
-// including its host and target-specific properties
-type Build struct {
-	CommonProps
-	BuildProps
-	Target TargetSpecific
-	Host   TargetSpecific
-	SplittableProps
-}
-
-func (l *Build) getTargetSpecific(tgt tgtType) *TargetSpecific {
-	if tgt == tgtTypeHost {
-		return &l.Host
-	} else if tgt == tgtTypeTarget {
-		return &l.Target
-	} else {
-		utils.Die("Unsupported target type: %s", tgt)
-	}
-	return nil
-}
-
-// These function check the boolean pointers - which are only filled if someone sets them
-// If not, the default value is returned
-
-func (l *Build) isHostSupported() bool {
-	if l.Host_supported == nil {
-		return false
-	}
-	return *l.Host_supported
-}
-
-func (l *Build) isTargetSupported() bool {
-	if l.Target_supported == nil {
-		return true
-	}
-	return *l.Target_supported
-}
-
-func (l *Build) isForwardingSharedLibrary() bool {
-	if l.Forwarding_shlib == nil {
-		return false
-	}
-	return *l.Forwarding_shlib
-}
-
-func (l *Build) isRpathWanted() bool {
-	if l.Add_lib_dirs_to_rpath == nil {
-		return false
-	}
-	return *l.Add_lib_dirs_to_rpath
-}
-
-func (l *Build) getBuildWrapperAndDeps(ctx blueprint.ModuleContext) (string, []string) {
-	if l.Build_wrapper != nil {
-		depargs := map[string]string{}
-		files := getDependentArgsAndFiles(ctx, depargs)
-
-		// Replace any property usage in buildWrapper
-		buildWrapper := *l.Build_wrapper
-		for k, v := range depargs {
-			buildWrapper = strings.Replace(buildWrapper, "${"+k+"}", v, -1)
-		}
-
-		return buildWrapper, files
-	}
-
-	return "", []string{}
-}
-
-// Add module paths to srcs, exclude_srcs, local_include_dirs, export_local_include_dirs
-// and post_install_tool
-func (l *BuildProps) processPaths(ctx blueprint.BaseModuleContext, g generatorBackend) {
-	prefix := projectModuleDir(ctx)
-
-	l.Export_local_include_dirs = utils.PrefixDirs(l.Export_local_include_dirs, prefix)
-	l.Export_local_system_include_dirs = utils.PrefixDirs(l.Export_local_system_include_dirs, prefix)
-
-	l.processBuildWrapper(ctx)
-}
-
-func (l *Build) processPaths(ctx blueprint.BaseModuleContext, g generatorBackend) {
-	l.BuildProps.processPaths(ctx, g)
-	l.CommonProps.processPaths(ctx, g)
-}
-
-func (c *CommonProps) processPaths(ctx blueprint.BaseModuleContext, g generatorBackend) {
-	prefix := projectModuleDir(ctx)
-
-	c.LegacySourceProps.processPaths(ctx, g)
-	c.InstallableProps.processPaths(ctx, g)
-	c.IncludeDirsProps.Local_include_dirs = utils.PrefixDirs(c.IncludeDirsProps.Local_include_dirs, prefix)
-}
 
 // library is a base class for modules which are generated from sets of object files
 type library struct {
@@ -700,153 +439,12 @@ func (l *library) exportLocalSystemIncludeDirs() []string {
 	return l.Properties.Export_local_system_include_dirs
 }
 
-type staticLibrary struct {
-	library
-}
-
-func (m *staticLibrary) GenerateBuildActions(ctx blueprint.ModuleContext) {
-	if isEnabled(m) {
-		getBackend(ctx).staticActions(m, ctx)
-	}
-}
-
-//// Support singleOutputModule
-
-func (m *staticLibrary) outputFileName() string {
-	return m.outputName() + ".a"
-}
-
-type sharedLibrary struct {
-	library
-	fileNameExtension string
-}
-
-// sharedLibrary supports:
-// * producing output using the linker
-// * producing a shared library
-// * stripping symbols from output
-var _ linkableModule = (*sharedLibrary)(nil)
-var _ sharedLibProducer = (*sharedLibrary)(nil)
-var _ stripable = (*sharedLibrary)(nil)
-
-func (m *sharedLibrary) getLinkName() string {
-	return m.outputName() + m.fileNameExtension
-}
-
-func (m *sharedLibrary) getSoname() string {
-	name := m.getLinkName()
-	if m.library.Properties.Library_version != "" {
-		var v = strings.Split(m.library.Properties.Library_version, ".")
-		name += "." + v[0]
-	}
-	return name
-}
-
-func (m *sharedLibrary) getRealName() string {
-	name := m.getLinkName()
-	if m.library.Properties.Library_version != "" {
-		name += "." + m.library.Properties.Library_version
-	}
-	return name
-}
-
-func (l *sharedLibrary) strip() bool {
-	return l.Properties.Strip != nil && *l.Properties.Strip
-}
-
-func (m *sharedLibrary) librarySymlinks(ctx blueprint.ModuleContext) map[string]string {
-	symlinks := map[string]string{}
-
-	if m.library.Properties.Library_version != "" {
-		// To build you need a symlink from the link name and soname.
-		// At runtime only the soname symlink is required.
-		soname := m.getSoname()
-		realName := m.getRealName()
-		if soname == realName {
-			utils.Die("module %s has invalid library_version '%s'",
-				m.Name(),
-				m.library.Properties.Library_version)
-		}
-		symlinks[m.getLinkName()] = soname
-		symlinks[soname] = realName
-	}
-
-	return symlinks
-}
-
-func (m *sharedLibrary) GenerateBuildActions(ctx blueprint.ModuleContext) {
-	if isEnabled(m) {
-		getBackend(ctx).sharedActions(m, ctx)
-	}
-}
-
-//// Support singleOutputModule
-
-func (m *sharedLibrary) outputFileName() string {
-	// Since we link against libraries using the library flag style,
-	// -lmod, return the name of the link library here rather than the
-	// real, versioned library.
-	return m.getLinkName()
-}
-
-//// Support sharedLibProducer
-
-func (m *sharedLibrary) getTocName() string {
-	return m.getRealName() + tocExt
-}
-
-type binary struct {
-	library
-}
-
-// binary supports:
-// * producing output using the linker
-// * stripping symbols from output
-var _ linkableModule = (*binary)(nil)
-var _ stripable = (*binary)(nil)
-
-func (l *binary) strip() bool {
-	return l.Properties.Strip != nil && *l.Properties.Strip
-}
-
-func (m *binary) GenerateBuildActions(ctx blueprint.ModuleContext) {
-	if isEnabled(m) {
-		getBackend(ctx).binaryActions(m, ctx)
-	}
-}
-
-//// Support singleOutputModule
-
-func (m *binary) outputFileName() string {
-	return m.outputName()
-}
-
 func (l *library) LibraryFactory(config *BobConfig, module blueprint.Module) (blueprint.Module, []interface{}) {
 	l.Properties.Features.Init(&config.Properties, CommonProps{}, BuildProps{}, SplittableProps{})
 	l.Properties.Host.init(&config.Properties, CommonProps{}, BuildProps{})
 	l.Properties.Target.init(&config.Properties, CommonProps{}, BuildProps{})
 
 	return module, []interface{}{&l.Properties, &l.SimpleName.Properties}
-}
-
-func staticLibraryFactory(config *BobConfig) (blueprint.Module, []interface{}) {
-	module := &staticLibrary{}
-	return module.LibraryFactory(config, module)
-}
-
-func sharedLibraryFactory(config *BobConfig) (blueprint.Module, []interface{}) {
-	module := &sharedLibrary{}
-	if config.Properties.GetBool("osx") {
-		module.fileNameExtension = ".dylib"
-	} else {
-		module.fileNameExtension = ".so"
-	}
-	return module.LibraryFactory(config, module)
-}
-
-func binaryFactory(config *BobConfig) (blueprint.Module, []interface{}) {
-	module := &binary{}
-	return module.LibraryFactory(config, module)
 }
 
 func getBinaryOrSharedLib(m blueprint.Module) (*library, bool) {
@@ -1032,140 +630,4 @@ func exportLibFlagsMutator(mctx blueprint.TopDownMutatorContext) {
 	})
 
 	checkForMultipleLinking(mctx.ModuleName(), allImportedStaticLibs, insideWholeLibs)
-}
-
-type graphMutatorHandler struct {
-	graphs map[tgtType]graph.Graph
-}
-
-const (
-	maxInt = int(^uint(0) >> 1)
-	minInt = -maxInt - 1
-)
-
-func (handler *graphMutatorHandler) ResolveDependencySortMutator(mctx blueprint.BottomUpMutatorContext) {
-	mainModule := mctx.Module()
-	if e, ok := mainModule.(enableable); ok {
-		if !isEnabled(e) {
-			return // Not enabled, so not needed
-		}
-	}
-	if _, ok := mainModule.(*defaults); ok {
-		return // ignore bob_defaults
-	}
-
-	mainModuleName := mainModule.Name()
-
-	if sp, ok := mainModule.(splittable); ok {
-		if sp.getTarget() != "" {
-			handler.graphs[sp.getTarget()].AddNode(mainModuleName)
-		}
-	}
-
-	var mainBuild *Build
-	if buildProps, ok := mainModule.(moduleWithBuildProps); ok {
-		mainBuild = buildProps.build()
-	} else {
-		return // ignore not a build
-	}
-
-	// This mutator is run after host/target splitting, so TargetType should have been set.
-	if !(mainBuild.TargetType == tgtTypeTarget || mainBuild.TargetType == tgtTypeHost) {
-		utils.Die("Cannot process dependencies on module '%s' with target type '%s'", mainModuleName, mainBuild.TargetType)
-	}
-
-	g := handler.graphs[mainBuild.TargetType]
-
-	for _, lib := range mainBuild.Static_libs {
-		if _, err := g.AddEdgeToExistingNodes(mainModuleName, lib); err != nil {
-			utils.Die("'%s' depends on '%s', but '%s' is either not defined or disabled", mainModuleName, lib, lib)
-		}
-		g.SetEdgeColor(mainModuleName, lib, "blue")
-	}
-
-	for _, lib := range mainBuild.Whole_static_libs {
-		if _, err := g.AddEdgeToExistingNodes(mainModuleName, lib); err != nil {
-			utils.Die("'%s' depends on '%s', but '%s' is either not defined or disabled", mainModuleName, lib, lib)
-		}
-		g.SetEdgeColor(mainModuleName, lib, "red")
-	}
-
-	temporaryPaths := map[string][]string{} // For preserving order in declaration
-
-	for i, previous := range mainBuild.Static_libs {
-		for j := i + 1; j < len(mainBuild.Static_libs); j++ {
-			lib := mainBuild.Static_libs[j]
-			if !g.IsReachable(lib, previous) {
-				if g.AddEdge(previous, lib) {
-					temporaryPaths[previous] = append(temporaryPaths[previous], lib)
-					g.SetEdgeColor(previous, lib, "pink")
-				}
-			}
-		}
-	}
-
-	sub := graph.GetSubgraph(g, mainModuleName)
-
-	// Remove temporary path
-	for key, list := range temporaryPaths {
-		for _, value := range list {
-			g.DeleteEdge(key, value)
-		}
-	}
-
-	// The order of static libraries influences performance by
-	// influencing memory layout. Where possible we want libraries
-	// that depend on each other to be as close as possible. Library
-	// order is determined by a topological sort.  Setting the
-	// priority changes the order that child nodes are visited.
-	//
-	// Libraries that are frequently called are more
-	// important and should be close to their callers. This information is not available in bob,
-	// so estimate this with the number of users.
-	//
-	// Libraries that are large, or will cause a large number of
-	// libraries to occur in the middle of the list, should be at the
-	// end of the list. Treat this as the cost of visiting the
-	// library. As an estimate of cost, count the number of libraries
-	// that would be pulled in.
-	//
-	// The node priority is calculated as 'A * importance - cost',
-	// where A is an arbitraty scaling factor.
-	//
-	// This is a bottom up mutator, so by the time we get to a binary
-	// (or shared library), this mutator will have run on all their
-	// dependencies and the (shared) graph will be complete (for the
-	// current module).
-	for _, nodeID := range sub.GetNodes() {
-		cost := graph.GetSubgraphNodeCount(sub, nodeID)
-		sources, _ := sub.GetSources(nodeID)
-		priority := len(sources)
-		sub.SetNodePriority(nodeID, (10*priority)-cost)
-	}
-
-	// The main library must always be evaluated first in the topological sort
-	sub.SetNodePriority(mainModuleName, minInt)
-
-	// We want those edges for calculating priority. After setting priority we can remove them.
-	sub.DeleteProxyEdges("red")
-
-	sub2 := graph.GetSubgraph(sub, mainModuleName)
-	sortedStaticLibs, isDAG := graph.TopologicalSort(sub2)
-
-	// Pop the module itself from the front of the list
-	sortedStaticLibs = sortedStaticLibs[1:]
-
-	if !isDAG {
-		utils.Die("We have detected cycle: %s", mainModuleName)
-	} else {
-		mainBuild.ResolvedStaticLibs = sortedStaticLibs
-	}
-
-	extraStaticLibsDependencies := utils.Difference(mainBuild.ResolvedStaticLibs, mainBuild.Static_libs)
-
-	mctx.AddVariationDependencies(nil, staticDepTag, extraStaticLibsDependencies...)
-
-	// This module may now depend on extra shared libraries, inherited from included
-	// static libraries. Add that dependency here.
-	mctx.AddVariationDependencies(nil, sharedDepTag, mainBuild.ExtraSharedLibs...)
 }
